Added element of randomness to the viewcount graph.
This prevents a small issue: when Brady's total was hidden (and thus the largest), it would ALWAYS be as wide as the window when revealed. This made the graph very predictable (when it shouldn't be.

This commit resolves the issue by making the largest bar between half the width of the window and the full width (the exact value is decided with Math.random) and scales all other values appropriately.